Implement security and config
Make sure the user has all privileges.
Use the Bucket that we used against Ddos to ensure that no one is sending too many requests to the server. Implement Client menu together with the server side (to avoid having 2 projects).
Parts of code is under construction till it starts working with the current project.. Client is taken from teacher and Mashreka.